Intermediate commit. Added some throws.
Currently implementing PCommand and POption and their usages. Instead
of having to parse every time a command is executed, all
commands will be parsed as they are registered. Main advantages are:
more orderly code, and users are notified if their annotations are
incorrect immediately when they register their commands instead of
when their commands are executed.